+++ src/main/java/org/opensearch/knn/index/query/nativelib/NativeEngineKnnVectorQuery.java
@@ -28,6 +28,7 @@ import org.opensearch.knn.index.query.ResultUtil;
 import org.opensearch.knn.index.query.rescore.RescoreContext;
 
 import java.io.IOException;
+import java.util.AbstractMap;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Comparator;
@@ -55,7 +56,7 @@ public class NativeEngineKnnVectorQuery extends Query {
         final IndexReader reader = indexSearcher.getIndexReader();
         final KNNWeight knnWeight = (KNNWeight) knnQuery.createWeight(indexSearcher, scoreMode, 1);
         List<LeafReaderContext> leafReaderContexts = reader.leaves();
-        List<Map<Integer, Float>> perLeafResults;
+        List<Map.Entry<LeafReaderContext, Map<Integer, Float>>> perLeafResults;
         RescoreContext rescoreContext = knnQuery.getRescoreContext();
         final int finalK = knnQuery.getK();
         if (rescoreContext == null) {
@@ -70,14 +71,15 @@ public class NativeEngineKnnVectorQuery extends Query {
             }
 
             StopWatch stopWatch = new StopWatch().start();
-            perLeafResults = doRescore(indexSearcher, leafReaderContexts, knnWeight, perLeafResults, finalK);
+            perLeafResults = doRescore(indexSearcher, knnWeight, perLeafResults, finalK);
             long rescoreTime = stopWatch.stop().totalTime().millis();
-            log.debug("Rescoring results took {} ms. oversampled k:{}, segments:{}", rescoreTime, firstPassK, leafReaderContexts.size());
+            log.debug("Rescoring results took {} ms. oversampled k:{}, segments:{}", rescoreTime, firstPassK, perLeafResults.size());
         }
         ResultUtil.reduceToTopK(perLeafResults, finalK);
         TopDocs[] topDocs = new TopDocs[perLeafResults.size()];
-        for (int i = 0; i < perLeafResults.size(); i++) {
-            topDocs[i] = ResultUtil.resultMapToTopDocs(perLeafResults.get(i), leafReaderContexts.get(i).docBase);
+        int i = 0;
+        for (Map.Entry<LeafReaderContext, Map<Integer, Float>> entry : perLeafResults) {
+            topDocs[i++] = ResultUtil.resultMapToTopDocs(entry.getValue(), entry.getKey().docBase);
         }
 
         TopDocs topK = TopDocs.merge(knnQuery.getK(), topDocs);
@@ -87,32 +89,29 @@ public class NativeEngineKnnVectorQuery extends Query {
         return createDocAndScoreQuery(reader, topK).createWeight(indexSearcher, scoreMode, boost);
     }
 
-    private List<Map<Integer, Float>> doSearch(
+    private List<Map.Entry<LeafReaderContext, Map<Integer, Float>>> doSearch(
         final IndexSearcher indexSearcher,
         List<LeafReaderContext> leafReaderContexts,
         KNNWeight knnWeight,
         int k
     ) throws IOException {
-        List<Callable<Map<Integer, Float>>> tasks = new ArrayList<>(leafReaderContexts.size());
+        List<Callable<Map.Entry<LeafReaderContext, Map<Integer, Float>>>> tasks = new ArrayList<>(leafReaderContexts.size());
         for (LeafReaderContext leafReaderContext : leafReaderContexts) {
             tasks.add(() -> searchLeaf(leafReaderContext, knnWeight, k));
         }
         return indexSearcher.getTaskExecutor().invokeAll(tasks);
     }
 
-    private List<Map<Integer, Float>> doRescore(
+    private List<Map.Entry<LeafReaderContext, Map<Integer, Float>>> doRescore(
         final IndexSearcher indexSearcher,
-        List<LeafReaderContext> leafReaderContexts,
         KNNWeight knnWeight,
-        List<Map<Integer, Float>> perLeafResults,
+        List<Map.Entry<LeafReaderContext, Map<Integer, Float>>> perLeafResults,
         int k
     ) throws IOException {
-        List<Callable<Map<Integer, Float>>> rescoreTasks = new ArrayList<>(leafReaderContexts.size());
-        for (int i = 0; i < perLeafResults.size(); i++) {
-            LeafReaderContext leafReaderContext = leafReaderContexts.get(i);
-            int finalI = i;
+        List<Callable<Map.Entry<LeafReaderContext, Map<Integer, Float>>>> rescoreTasks = new ArrayList<>(perLeafResults.size());
+        for (Map.Entry<LeafReaderContext, Map<Integer, Float>> entry : perLeafResults) {
             rescoreTasks.add(() -> {
-                BitSet convertedBitSet = ResultUtil.resultMapToMatchBitSet(perLeafResults.get(finalI));
+                BitSet convertedBitSet = ResultUtil.resultMapToMatchBitSet(entry.getValue());
                 final ExactSearcher.ExactSearcherContext exactSearcherContext = ExactSearcher.ExactSearcherContext.builder()
                     .matchedDocs(convertedBitSet)
                     // setting to false because in re-scoring we want to do exact search on full precision vectors
@@ -121,7 +120,8 @@ public class NativeEngineKnnVectorQuery extends Query {
                     .isParentHits(false)
                     .knnQuery(knnQuery)
                     .build();
-                return knnWeight.exactSearch(leafReaderContext, exactSearcherContext);
+                final Map<Integer, Float> docIdScoreMap = knnWeight.exactSearch(entry.getKey(), exactSearcherContext);
+                return new AbstractMap.SimpleEntry<>(entry.getKey(), docIdScoreMap);
             });
         }
         return indexSearcher.getTaskExecutor().invokeAll(rescoreTasks);
@@ -158,13 +158,14 @@ public class NativeEngineKnnVectorQuery extends Query {
         return starts;
     }
 
-    private Map<Integer, Float> searchLeaf(LeafReaderContext ctx, KNNWeight queryWeight, int k) throws IOException {
+    private Map.Entry<LeafReaderContext, Map<Integer, Float>> searchLeaf(LeafReaderContext ctx, KNNWeight queryWeight, int k)
+        throws IOException {
         final Map<Integer, Float> leafDocScores = queryWeight.searchLeaf(ctx, k);
         final Bits liveDocs = ctx.reader().getLiveDocs();
         if (liveDocs != null) {
             leafDocScores.entrySet().removeIf(entry -> liveDocs.get(entry.getKey()) == false);
         }
-        return leafDocScores;
+        return new AbstractMap.SimpleEntry<>(ctx, leafDocScores);
     }
 
     @Override